proper tracking of skipped downloads, use max available parallelism
